--- src/test/java/nablarch/common/authorization/action/BarAction.java
@@ -1,9 +1,0 @@
-package nablarch.common.authorization.action;
-
-import nablarch.common.authorization.CheckAuthority;
-
-public class BarAction {
-
-    @CheckAuthority("FOO")
-    public void publicMethod() {}
-}
