--- src/test/java/nablarch/common/authorization/action/ParentAction.java
@@ -1,9 +1,0 @@
-package nablarch.common.authorization.action;
-
-import nablarch.common.authorization.CheckAuthority;
-
-public class ParentAction {
-
-    @CheckAuthority("FOO")
-    public void parentMethod() {}
-}
