--- src/test/java/nablarch/common/authorization/action/sub/FizzAction.java
@@ -1,9 +1,0 @@
-package nablarch.common.authorization.action.sub;
-
-import nablarch.common.authorization.CheckAuthority;
-
-public class FizzAction {
-
-    @CheckAuthority("FOO")
-    public void publicMethod() {}
-}
