--- src/test/java/nablarch/common/authorization/action/sub/FizzService.java
@@ -1,9 +1,0 @@
-package nablarch.common.authorization.action.sub;
-
-import nablarch.common.authorization.CheckAuthority;
-
-public class FizzService {
-
-    @CheckAuthority("FOO")
-    public void publicMethod() {}
-}
