+++ src/test/java/nablarch/common/authorization/role/action/BarAction.java
@@ -1,0 +1,9 @@
+package nablarch.common.authorization.role.action;
+
+import nablarch.common.authorization.role.CheckAuthority;
+
+public class BarAction {
+
+    @CheckAuthority("FOO")
+    public void publicMethod() {}
+}
