+++ src/test/java/nablarch/common/authorization/role/action/FooService.java
@@ -1,0 +1,9 @@
+package nablarch.common.authorization.role.action;
+
+import nablarch.common.authorization.role.CheckAuthority;
+
+public class FooService {
+
+    @CheckAuthority("FOO")
+    public void publicMethod() {}
+}
