+++ src/test/java/nablarch/common/authorization/role/action/ParentAction.java
@@ -1,0 +1,9 @@
+package nablarch.common.authorization.role.action;
+
+import nablarch.common.authorization.role.CheckAuthority;
+
+public class ParentAction {
+
+    @CheckAuthority("FOO")
+    public void parentMethod() {}
+}
