+++ src/test/java/nablarch/common/authorization/role/action/sub/FizzAction.java
@@ -1,0 +1,9 @@
+package nablarch.common.authorization.role.action.sub;
+
+import nablarch.common.authorization.role.CheckAuthority;
+
+public class FizzAction {
+
+    @CheckAuthority("FOO")
+    public void publicMethod() {}
+}
