+++ src/test/java/nablarch/common/authorization/role/action/sub/FizzService.java
@@ -1,0 +1,9 @@
+package nablarch.common.authorization.role.action.sub;
+
+import nablarch.common.authorization.role.CheckAuthority;
+
+public class FizzService {
+
+    @CheckAuthority("FOO")
+    public void publicMethod() {}
+}
